+++ src/lib.rs
@@ -1,3 +1,6 @@
 pub mod list_ordered_multimap;
 
 pub use self::list_ordered_multimap::ListOrderedMultimap;
+
+#[cfg(feature = "serde-1")]
+mod serde;