+++ src/list_ordered_multimap.rs
@@ -1,5 +1,6 @@
 use dlv_list::{
-    Drain as VecListDrain, Index, Iter as VecListIter, IterMut as VecListIterMut, VecList,
+    Drain as VecListDrain, Index, IntoIter as VecListIntoIter, Iter as VecListIter,
+    IterMut as VecListIterMut, VecList,
 };
 use hashbrown::hash_map::{RawEntryMut, RawOccupiedEntryMut};
 use hashbrown::HashMap;
@@ -1669,6 +1670,22 @@ where
     }
 }
 
+impl<Key, Value, State> IntoIterator for ListOrderedMultimap<Key, Value, State>
+where
+    Key: Eq + Hash + Clone,
+    State: BuildHasher,
+{
+    type IntoIter = IntoIter<Key, Value>;
+    type Item = (Key, Value);
+
+    fn into_iter(self) -> Self::IntoIter {
+        IntoIter {
+            keys: self.keys,
+            iter: self.values.into_iter(),
+        }
+    }
+}
+
 impl<Key, Value, State> PartialEq for ListOrderedMultimap<Key, Value, State>
 where
     Key: Eq + Hash,
@@ -3014,6 +3031,70 @@ impl<'map, Key, Value> Iterator for IterMut<'map, Key, Value> {
     }
 }
 
+/// An iterator that owns and yields all key-value pairs in a multimap. The order of
+/// the yielded items is always in the order that they were inserted.
+pub struct IntoIter<Key, Value> {
+    // The list of the keys in the map. This is ordered by time of insertion.
+    keys: VecList<Key>,
+
+    /// The iterator over the list of all values. This is ordered by time of insertion.
+    iter: VecListIntoIter<ValueEntry<Key, Value>>,
+}
+
+impl<Key, Value> IntoIter<Key, Value> {
+    /// Creates an iterator that yields immutable references to all key-value pairs in a multimap.
+    pub fn iter(&self) -> Iter<Key, Value> {
+        Iter {
+            keys: &self.keys,
+            iter: self.iter.iter(),
+        }
+    }
+}
+
+impl<Key, Value> Debug for IntoIter<Key, Value>
+where
+    Key: Debug,
+    Value: Debug,
+{
+    fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
+        formatter.write_str("IntoIter(")?;
+        formatter.debug_list().entries(self.iter()).finish()?;
+        formatter.write_str(")")
+    }
+}
+
+impl<Key, Value> DoubleEndedIterator for IntoIter<Key, Value>
+where
+    Key: Clone,
+{
+    fn next_back(&mut self) -> Option<Self::Item> {
+        let value_entry = self.iter.next_back()?;
+        let key = self.keys.get(value_entry.key_index).cloned().unwrap();
+        Some((key, value_entry.value))
+    }
+}
+
+impl<Key, Value> ExactSizeIterator for IntoIter<Key, Value> where Key: Clone {}
+
+impl<Key, Value> FusedIterator for IntoIter<Key, Value> where Key: Clone {}
+
+impl<Key, Value> Iterator for IntoIter<Key, Value>
+where
+    Key: Clone,
+{
+    type Item = (Key, Value);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let value_entry = self.iter.next()?;
+        let key = self.keys.get(value_entry.key_index).cloned().unwrap();
+        Some((key, value_entry.value))
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.iter.size_hint()
+    }
+}
+
 /// An iterator that yields immutable references to all keys and their value iterators. The order of
 /// the yielded items is always in the order the keys were first inserted.
 pub struct KeyValues<'map, Key, Value, State = RandomState> {
@@ -3706,6 +3787,7 @@ mod test {
         check_bounds::<EntryValuesMut<'static, (), ()>>();
         check_bounds::<Iter<'static, (), ()>>();
         check_bounds::<IterMut<'static, (), ()>>();
+        check_bounds::<IntoIter<(), ()>>();
         check_bounds::<KeyValues<'static, (), ()>>();
         check_bounds::<KeyValuesDrain<'static, (), ()>>();
         check_bounds::<KeyValuesEntryDrain<'static, (), ()>>();
@@ -4085,6 +4167,27 @@ mod test {
     }
 
     #[test]
+    fn test_iter_size_hint() {
+        let mut map = ListOrderedMultimap::new();
+
+        map.insert("key1", "value1");
+        map.append("key2", "value2");
+        map.append("key2", "value3");
+        map.append("key1", "value4");
+
+        let mut iter = map.iter();
+        assert_eq!(iter.size_hint(), (4, Some(4)));
+        iter.next();
+        assert_eq!(iter.size_hint(), (3, Some(3)));
+        iter.next();
+        assert_eq!(iter.size_hint(), (2, Some(2)));
+        iter.next();
+        assert_eq!(iter.size_hint(), (1, Some(1)));
+        iter.next();
+        assert_eq!(iter.size_hint(), (0, Some(0)));
+    }
+
+    #[test]
     fn test_iter_mut_debug() {
         let mut map = ListOrderedMultimap::new();
 
@@ -4162,7 +4265,7 @@ mod test {
     }
 
     #[test]
-    fn test_iter_size_hint() {
+    fn test_into_iter_debug() {
         let mut map = ListOrderedMultimap::new();
 
         map.insert("key1", "value1");
@@ -4170,7 +4273,63 @@ mod test {
         map.append("key2", "value3");
         map.append("key1", "value4");
 
-        let mut iter = map.iter();
+        let iter = map.into_iter();
+        assert_eq!(
+            format!("{:?}", iter),
+            r#"IntoIter([("key1", "value1"), ("key2", "value2"), ("key2", "value3"), ("key1", "value4")])"#
+        );
+    }
+
+    #[test]
+    fn test_into_iter_double_ended() {
+        let mut map = ListOrderedMultimap::new();
+
+        map.insert("key1", "value1");
+        map.append("key2", "value2");
+        map.append("key2", "value3");
+        map.append("key1", "value4");
+
+        let mut iter = map.into_iter();
+        assert_eq!(iter.next(), Some(("key1", "value1")));
+        assert_eq!(iter.next_back(), Some(("key1", "value4")));
+        assert_eq!(iter.next(), Some(("key2", "value2")));
+        assert_eq!(iter.next_back(), Some(("key2", "value3")));
+        assert_eq!(iter.next(), None);
+    }
+
+    #[test]
+    fn test_into_iter_empty() {
+        let map: ListOrderedMultimap<&str, &str> = ListOrderedMultimap::new();
+        let mut iter = map.into_iter();
+        assert_eq!(iter.next_back(), None);
+        assert_eq!(iter.next(), None);
+    }
+
+    #[test]
+    fn test_into_iter_fused() {
+        let mut map = ListOrderedMultimap::new();
+
+        map.insert("key", "value");
+
+        let mut iter = map.into_iter();
+        assert_eq!(iter.next(), Some(("key", "value")));
+        assert_eq!(iter.next(), None);
+        assert_eq!(iter.next_back(), None);
+        assert_eq!(iter.next(), None);
+        assert_eq!(iter.next_back(), None);
+        assert_eq!(iter.next(), None);
+    }
+
+    #[test]
+    fn test_into_iter_size_hint() {
+        let mut map = ListOrderedMultimap::new();
+
+        map.insert("key1", "value1");
+        map.append("key2", "value2");
+        map.append("key2", "value3");
+        map.append("key1", "value4");
+
+        let mut iter = map.into_iter();
         assert_eq!(iter.size_hint(), (4, Some(4)));
         iter.next();
         assert_eq!(iter.size_hint(), (3, Some(3)));